+++ src/core/bk_ut.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING
 from PyQt6.QtCore import (QModelIndex, pyqtSlot, QPoint, QThread,
     QTimer, QAbstractTableModel, Qt,
 )
-from PyQt6.QtGui import QResizeEvent, QKeySequence, QShortcut
+from PyQt6.QtGui import QResizeEvent, QKeySequence, QShortcut, QAction
 from PyQt6.QtWidgets import (QMenu, QTreeView, QHeaderView,
     QMessageBox, QApplication,
 )
@@ -134,6 +134,7 @@ def short_delete_folder():
 
 @pyqtSlot()
 def show_main_menu():
+    is_db_opened = bool(ag.db.conn)
     menu = QMenu(self)
     if not ag.single_instance:
         menu.addAction('New window')
@@ -141,10 +142,16 @@ def show_main_menu():
     menu.addAction('Create/Open DB')
     menu.addAction('Select DB from list')
     menu.addSeparator()
-    menu.addAction('Scan disk for files')
+    act_scan = QAction('Scan disk for files')
+    act_scan.setEnabled(is_db_opened)
+    menu.addAction(act_scan)
     menu.addSeparator()
-    menu.addAction('Report duplicate files')
-    menu.addAction('Report files with same names')
+    act_dup = QAction('Report duplicate files')
+    act_dup.setEnabled(is_db_opened)
+    menu.addAction(act_dup)
+    act_same = QAction('Report files with same names')
+    act_same.setEnabled(is_db_opened)
+    menu.addAction(act_same)
     menu.addSeparator()
     menu.addAction('Preferences')
     menu.addSeparator()
@@ -192,8 +199,6 @@ def file_list_resize_0(e: QResizeEvent):
     hdr = ag.file_list.header()
     ag.file_list.resizeEvent = file_list_resize
     ag.signals_.toggle_column.connect(change_sum_width)
-    file_list_resize(e)
-    ag.signals_.app_mode_changed.emit(ag.appMode.NIL.value)
 
 def restore_dirs():
     low_bk.set_dir_model()
@@ -214,7 +219,7 @@ def restore_dirs():
 
 def header_restore(model: QAbstractTableModel):
     global sum_width, min_width
-    hdr = ag.file_list.header()
+    hdr: QHeaderView = ag.file_list.header()
     try:
         state = ag.get_setting("FILE_LIST_HEADER")
         if state:
@@ -237,6 +242,22 @@ def header_restore(model: QAbstractTableModel):
 
     hdr.setSectionResizeMode(QHeaderView.ResizeMode.Interactive)
     hdr.sectionResized.connect(resized_column)
+
+    hdr.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
+    hdr.customContextMenuRequested.connect(header_menu)
+
+@pyqtSlot(QPoint)
+def header_menu(pos: QPoint):
+    hdr: QHeaderView = ag.file_list.header()
+    idx = hdr.logicalIndexAt(pos)
+    if idx:
+        me = QMenu()
+        me.addAction(f'Hide column "{low_bk.fields[idx]}"')
+        action = me.exec(hdr.mapToGlobal(
+            QPoint(pos.x(), pos.y() + hdr.height()))
+        )
+        if action:
+            low_bk.toggle_show_column(False, idx)
 
 @pyqtSlot(int, int, int)
 def resized_column(localIdx: int, oldSize: int, newSize: int):
@@ -308,10 +329,6 @@ def msg_copy_menu(pos: QPoint):
     menu.addAction("Copy message")
     action = menu.exec(ag.app.ui.msg.mapToGlobal(pos))
     if action:
-        copy_message()
-
-def copy_message():
-    if ag.app.ui.msg.text():
         QApplication.clipboard().setText(ag.app.ui.msg.text())
         ag.app.ui.msg.clear()