+++ src/core/utils.py
@@ -31,13 +31,10 @@ def show_message_box(
 
     return dlg.exec()
 
-def get_log_file_name() -> str:
-    from datetime import datetime as dt
+def get_log_path() -> str:
     log_path = tug.get_app_setting("DEFAULT_LOG_PATH", "")
     r_path = Path(log_path) if log_path else Path().resolve()
-    file_name = f"{dt.now():%b %d %H.%M.%S}.log"
-    file = r_path / file_name
-    return file.as_posix()
+    return r_path.as_posix()
 
 def set_logger():
     logger.remove()
@@ -47,8 +44,8 @@ def set_logger():
 
     fmt = "{time:%y-%b-%d %H:%M:%S} | {level} | {module}.{function}({line}): {message}"
 
-    LOG_FILE = get_log_file_name()
-    logger.add(LOG_FILE, format=fmt, enqueue=True)
-    # logger.add(sys.stderr, format=fmt, enqueue=True)
+    log_path = get_log_path()
+    logger.add(log_path+"/fileo.log", format=fmt, rotation="1 days", retention=3)
+    # logger.add(sys.stderr,  format='"{file.path}", line {line}, {function} - {message}')
+    logger.info(f"START =================> {log_path}")
     logger.info(f'{ag.app_name()=}, {ag.app_version()=}')
-    logger.info(f"START =================> {LOG_FILE}")